Made start-all-factors.sh interactive, for selectively restarting data sources.

+++ start-all-factors.sh
@@ -2,14 +2,24 @@
 
 cd factors
 for script in *.py *.sh ; do
+  STARTTHIS=y
   pid=$(ps axuwww | egrep -v 'grep|vim' | grep "$script" | awk '{print $2}')
   if [ "$pid" != "" ]; then
+    STARTTHIS=n
     for p in "$pid" ; do
       echo kill "$pid"
-      ps axuwww | grep -v grep | grep " $pid "
-      kill "$pid"
+      ps axuwww | grep -v grep | grep " $pid  "
+      # TODO: ask user whether to kill this process
+      echo -n 'Kill this? [y/N] '
+      read ANS
+      if [ "$ANS" = 'y' ]; then
+        kill "$pid"
+        STARTTHIS=y
+      fi
     done
   fi
-  echo "Starting $script ..."
-  "./$script" &
+  if [ "$STARTTHIS" = 'y' ]; then
+    echo "Starting $script ..."
+    "./$script" &
+  fi
 done
